Adapts tests to reflect different default-timezone
This Commit adapts the current test to include different default
server-timezones. Before this the timezone was set to UTC in the
phpunit.xml.dist which might not be the case in different
server-configurations. Therefore I've added a way to set the
default-servertimezone to different values before running the
filter-tests to take those different timezones into account. The tests
should ALWAYS return the same time or timestamp for a given
datetime-value.